Documented users API with Swagger. Fixed delete current user and delete user by ID Cloudinary issue

+++ src/controllers/v1/user/delete_current_user.ts
@@ -20,10 +20,12 @@ const deleteCurrentUser = async (
       .exec();
     const publicIds = blogs.map(({ banner }) => banner.publicId);
 
-    await cloudinary.api.delete_resources(publicIds);
-    logger.info("Multiple banner images deleted from Cloudinary", {
-      publicIds,
-    });
+    if (publicIds.length > 0) {
+      await cloudinary.api.delete_resources(publicIds);
+      logger.info("Multiple banner images deleted from Cloudinary", {
+        publicIds,
+      });
+    }
 
     await Blog.deleteMany({ author: userId });
     logger.info("All blogs by user deleted", { userId, blogs });
